google_places_searcher/review_spider/requirements.txt
scrapy==1.8.0
psycopg2-binary==2.8.4
pymongo==3.9.0
scrapy-splash===0.7.2
schedule==0.6.0
